Project-SlugShield-IDS/requirements.txt
scapy==2.6.1
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-dotenv==1.0.0
pytest==8.4.2
PyYAML==6.0.1